requirements.txt

reflex==0.8.14
PyPDF2
elevenlabs
google-cloud-texttospeech
PyMuPDF
google-generativeai
google-cloud-aiplatform
httpx